Fix the forward declaration issue of the todolist::Todo
This commit fix an issue related to the compilation issue of forward
declaration for the todolist::Todo. The djinni generates the todo.hpp
and todo_list.hpp files. The todo_list.hpp has a forward declaration of
the Todo struct. The code in todo_list.hpp before this commit was
including the file todo.hpp, but since it is an autogenerated file, it
is not recommended to do that. Instead, this commit includes the
todo.hpp in the todo_list_impl.hpp. Please, not that this requires an
update in the content for the file todo_list_impl.hpp in the tutorial
available in [1].

In addition, this commit adds an entry in the Makefile for cleaning the
files generated in the compilation process.

[1]
http://mobilecpptutorials.com/todo-app-using-djinni-and-sqlite-part-1-cplusplus/